requirements.txt
pytest==7.2.0
allure-pytest==2.12.0
requests~=2.28.1
jsonschema==4.17.3
pytest-xdist==3.0.2
curlify~=2.2.1
selene>=2.0.0b16
pydantic>=1.10.2
python-dotenv>=0.21.0